+++ BlogMvcCore/Startup.cs
@@ -62,15 +62,8 @@ namespace BlogMvcCore
             app.UseEndpoints(endpoints =>
             {
                 endpoints.MapControllerRoute(
-                    name: "Authentication",
+                    name: "default",
                     pattern: "{controller=Authentication}/{action=index}/{id?}");
-
-                endpoints.MapControllerRoute(
-                    name: "User",
-                    pattern: "{controller=User}/{action=Index}/{id?}");
-                endpoints.MapControllerRoute(
-                    name: "Post",
-                    pattern: "{controller=Post}/{action=index}/{id?}");
             });
         }
     }
