+++ BlogMvcCore/Views/Comment/EditComment.cshtml
@@ -9,7 +9,7 @@
         </tr>
         <tr>
             <td>
-                @using (Html.BeginForm("UpdateComment", "User", FormMethod.Post))
+                @using (Html.BeginForm("UpdateComment", "Comment", FormMethod.Post))
                 {
                     @Html.TextBox("text", $"{Model.Text}", htmlAttributes: true)
                     <input type="hidden" id="commentID" name="commentID" value=@Model.ID />
