+++ BlogMvcCore/Views/User/VisitUserPage.cshtml
@@ -6,6 +6,6 @@
 <div>
     <table>
         <tr><td>@await Html.PartialAsync("_UserInfo")</td></tr>
-        <tr><td>@Html.Partial("~/Post/_UserPost.cshtml")</td></tr>
+        <tr><td>@Html.Partial("~/Views/Post/_UserPost.cshtml")</td></tr>
     </table>
 </div>
